pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "translation_game"
version = "0.1.0"
description = "A side project for translation game"
authors = ["Hyun Jae Moon <calhyunjaemoon@gmail.com>"]

[tool.poetry.dependencies]
google-generativeai = "0.6.0"
python = "^3.9"
python-dotenv = "^0.19.0"
requests = "^2.25.1"
ipython = "^7.29.0"
pygame = "^2.0.1"
pygbag = "^0.9.1"
